requirements.txt
Django>=1.2
Scrapy>=0.14
django-celery
django-kombu
PIL

